move http2 to http2.test.js
